requirements.txt
numpy==1.26.4
uvicorn==0.30.6
pandas==2.2.3
pillow==10.4.0
tensorflow==2.17.0
python-dotenv==1.0.1
joblib==1.4.2
typing_extensions==4.12.2
scikit-learn==1.5.2
# pydantic==2.9.2
# protobuf==5.28.2
fastapi[standard]
google-generativeai==0.8.1
